Move WakeComputeBackend trait to proxy::wake_compute
